@@ -1,12 +1,15 @@
-package com.udacity.sandwichclub;
+package com.udacity.sandwichclub.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.udacity.sandwichclub.R;
 
 public class MainActivity extends AppCompatActivity
 {
